🚚 Move test files from field/ to fieldname/ and modifier/
