+++ tests/sigmahq/fieldname/test_sigmahq_invalid_fieldname_validator.py
@@ -1,7 +1,7 @@
 from sigma.correlations import SigmaCorrelationRule
 from sigma.rule import SigmaRule
 
-from sigma.validators.sigmahq.field import (
+from sigma.validators.sigmahq.fieldname import (
     SigmahqInvalidFieldnameIssue,
     SigmahqInvalidFieldnameValidator,
 )
